+++ packages/react-deprecate-warning/src/index.tsx
@@ -1,0 +1,2 @@
+export * from "./WithDeprecatedComponent"
+export * from "./WithDeprecatedProp"