+++ draft-packages/loading-placeholder/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.64](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.3.63...@kaizen/draft-loading-placeholder@1.3.64) (2020-12-01)
+
+**Note:** Version bump only for package @kaizen/draft-loading-placeholder
+
+
+
+
+
 ## [1.3.63](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.3.62...@kaizen/draft-loading-placeholder@1.3.63) (2020-11-25)
 
 **Note:** Version bump only for package @kaizen/draft-loading-placeholder