+++ draft-packages/table/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.5...@kaizen/draft-table@3.2.6) (2020-12-01)
+
+
+### Bug Fixes
+
+* Added missing dependencies. Updated eslint. ([#875](https://github.com/cultureamp/kaizen-design-system/issues/875)) ([182a9ca](https://github.com/cultureamp/kaizen-design-system/commit/182a9cafa9cf9795dcdd936cfef7bac432d3c28f))
+
+
+
+
+
+## [3.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.4...@kaizen/draft-table@3.2.5) (2020-12-01)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [3.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.3...@kaizen/draft-table@3.2.4) (2020-11-30)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
 ## [3.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.2...@kaizen/draft-table@3.2.3) (2020-11-26)