+++ legacy-packages/menu-list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.49](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.5.48...@kaizen/draft-menu-list@1.5.49) (2020-12-01)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
 ## [1.5.48](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.5.47...@kaizen/draft-menu-list@1.5.48) (2020-11-25)
 
 **Note:** Version bump only for package @kaizen/draft-menu-list